+++ js/main.js
@@ -1,5 +1,6 @@
 'use strict';
 
+import './api.js'
+import './map.js';
 import './popup.js';
 import './form.js';
-import './map.js';
